[future] Exception package move to common module
