+++ streampark-common/src/main/java/org/apache/streampark/common/domain/ResponseCode.java
@@ -15,20 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.streampark.console.base.exception;
+package org.apache.streampark.common.domain;
 
-import org.apache.streampark.common.util.ExceptionUtils;
+public interface ResponseCode {
 
-public class AlertException extends ApiAlertException {
-  public AlertException(String message) {
-    super(message);
-  }
+  Long CODE_SUCCESS = 200L;
 
-  public AlertException(Throwable cause) {
-    super(ExceptionUtils.stringifyException(cause));
-  }
+  Long CODE_FAIL = 500L;
 
-  public AlertException(String message, Throwable cause) {
-    super(message, cause);
-  }
+  Long CODE_FORBIDDEN = 403L;
+
+  Long CODE_FAIL_ALERT = 501L;
+
+  Long CODE_FAIL_DETAIL = 502L;
 }
