+++ streampark-common/src/main/java/org/apache/streampark/common/exception/ApplicationException.java
@@ -15,17 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.streampark.console.base.domain;
+package org.apache.streampark.common.exception;
 
-public interface ResponseCode {
+/** Applies to all application exceptions */
+public class ApplicationException extends ApiAlertException {
+  public ApplicationException(String message) {
+    super(message);
+  }
 
-  Long CODE_SUCCESS = 200L;
-
-  Long CODE_FAIL = 500L;
-
-  Long CODE_FORBIDDEN = 403L;
-
-  Long CODE_FAIL_ALERT = 501L;
-
-  Long CODE_FAIL_DETAIL = 502L;
+  public ApplicationException(Throwable cause) {
+    super(cause.getMessage());
+  }
 }
