+++ streampark-common/src/main/java/org/apache/streampark/common/exception/IllegalFileTypeException.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.streampark.console.base.exception;
+package org.apache.streampark.common.exception;
 
 /** This exception is thrown when there is an error in the file type */
 public class IllegalFileTypeException extends ApiAlertException {
