+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/base/domain/RestResponse.java
@@ -17,6 +17,8 @@
 
 package org.apache.streampark.console.base.domain;
 
+import org.apache.streampark.common.domain.ResponseCode;
+
 import org.slf4j.helpers.MessageFormatter;
 
 import javax.validation.constraints.NotNull;
