+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/service/alert/AlertNotifyService.java
@@ -17,7 +17,7 @@
 
 package org.apache.streampark.console.core.service.alert;
 
-import org.apache.streampark.console.base.exception.AlertException;
+import org.apache.streampark.common.exception.AlertException;
 import org.apache.streampark.console.core.bean.AlertConfigParams;
 import org.apache.streampark.console.core.bean.AlertTemplate;
