+++ streampark-console/streampark-console-service/src/test/java/org/apache/streampark/console/core/utils/YarnQueueLabelExpressionTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.streampark.console.core.utils;
 
-import org.apache.streampark.console.base.exception.ApiAlertException;
+import org.apache.streampark.common.exception.ApiAlertException;
 
 import org.junit.jupiter.api.Test;
